Reorganise tests
